requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
pyzbar==0.1.9
Pillow==10.1.0
pandas==2.1.4
lxml==5.1.0
python-dotenv==1.0.0
openpyxl==3.1.2
pytest==7.4.3